+++ src/cset.c
@@ -130,7 +130,7 @@ usage:		sprintf(buf, "bk help %s", av[0]);
 	}
 
 	if (streq(av[0], "makepatch")) copts.makepatch = 1;
-	copts.notty = !isatty(1);
+	copts.notty = (getenv("BK_NOTTY") != 0);
 
 	while (
 	    (c =
@@ -651,7 +651,7 @@ csetlist(cset_t *cs, sccs *cset)
 			unlink(cat);
 			goto fail;
 		}
-		if (sysio(cat, csort, 0, "bk", "_keysort", SYS)) {
+		if (sysio(cat, csort, 0, "bk", "_sort", SYS)) {
 			unlink(cat);
 			goto fail;
 		}
@@ -973,7 +973,7 @@ add(FILE *diffs, char *buf)
 
 	p = basenm(buf);
 	*p = 'd';
-	unlink(buf); /* remove d.file */
+	if (d == sccs_top(s)) unlink(buf); /* remove d.file */
 
 	sccs_sdelta(s, sccs_ino(s), buf);
 	fprintf(diffs, "> %s ", buf);
